// DXC_Frontend/src/api/models/questionDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */
import type { AnswerDto } from './answerDto'

export interface QuestionDto {
  id?: number
  /** @nullable */
  noiDung?: string | null
  /** @nullable */
  cauHoiTuLuan?: string | null
  /** @nullable */
  stt?: number | null
  /** @nullable */
  answers?: AnswerDto[] | null
}